+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - A new CLI flag `--debug` causes `conjur-preflight` to log more verbose
   information about the execution of the application and its checks.
   [conjurinc/conjur-preflight#19](https://github.com/conjurinc/conjur-preflight/pull/19)
+- `conjur-preflight` now includes disk related checks for read, write, and sync
+  latency, as well as read and write operations per second (IOPs). These require
+  `fio` and `libaio` to be present as a prerequesite for these checks.
+  [conjurinc/conjur-preflight#19](https://github.com/conjurinc/conjur-preflight/pull/19)
 
 ### Fixed
 - Previously, the application version was not properly embedded in the final